@snajpa 2022-09-21
I've been experimenting with this approach and for us it wasn't enough - there's also the `dsl_dataset_check_quota` function, which has probably made even more noticeable difference - after treating it accordingly, inspired by this patch.

We've settled on `ext_quota` to amount to `zfs_dirty_data_max * zfs_dirty_data_sync_percent / 100` when quota is set; that way we can be sure txgs are closed as they should even when there are datasets near quota.

In our case that means max overshoot of 6.4 GB max, tolerable for our use-case (OS-level containers)